requirements.txt
# Core Agent Framework
agent-framework-azure-ai

# Azure SDKs
azure-identity
azure-monitor-opentelemetry

# API Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
aiosqlite==0.20.0

# Scheduling
apscheduler==3.10.4

# Monitoring & Logging
prometheus-client==0.21.0
python-json-logger==3.2.0
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-instrumentation-fastapi==0.49b2

# HTTP Client
httpx==0.27.2
aiohttp==3.11.7

# Configuration
python-dotenv==1.0.1
pyyaml==6.0.2

# Utilities
python-dateutil==2.9.0.post0
pytz==2024.2

# Development & Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
black==24.10.0
ruff==0.7.4
mypy==1.13.0
